requirements.txt
pymodbus==3.6.3
paho-mqtt==1.6.1
